requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
Django==3.2
django-appconf==1.0.4
django-imagekit==4.0.2
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pilkit==2.0
Pillow==8.2.0
pycodestyle==2.7.0
pylint==2.8.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
